pyproject.toml
[tool.poetry]
name = "news-email-sender"
version = "0.1.0"
description = ""
authors = ["Macro1027 <marcoleepython@gmail.com>"]
readme = "README.md"
packages = [{include = "news_email_sender"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
